backend/requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
drf-nested-routers